requirements.txt

argon2-cffi==21.3.0
celery==5.2.7
django-configurations==2.4
Django==3.2.7
django-cors-headers==3.13.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-filter==22.1
django-redis==5.2.0
django-timezone-field==5.0
firebase-admin==5.2.0
graphene-django==2.15.0
gunicorn==20.1.0
httplib2==0.20.2
hypothesis==6.61.0
inflection==0.5.1
ipython==8.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2021.3
twilio==7.14.2
psycopg2==2.9.3
django-graphene-permissions==0.0.4
stripe==4.2.0
